PCT-625: move all api calls to separate package
