+++ bin/percona-agent-installer/installer/config.go
@@ -24,165 +24,15 @@ import (
 	"github.com/percona/percona-agent/data"
 	"github.com/percona/percona-agent/instance"
 	pctLog "github.com/percona/percona-agent/log"
-	mmMySQL "github.com/percona/percona-agent/mm/mysql"
-	mmServer "github.com/percona/percona-agent/mm/system"
 	"github.com/percona/percona-agent/pct"
-	"github.com/percona/percona-agent/qan"
-	sysconfigMySQL "github.com/percona/percona-agent/sysconfig/mysql"
-	"log"
-	"net/http"
 )
-
-func (i *Installer) getMmServerConfig(si *proto.ServerInstance) (*proto.AgentConfig, error) {
-	url := pct.URL(i.agentConfig.ApiHostname, "/configs/mm/default-server")
-	code, data, err := i.api.Get(i.agentConfig.ApiKey, url)
-	if i.flags.Bool["debug"] {
-		log.Printf("code=%d\n", code)
-		log.Printf("err=%s\n", err)
-	}
-	if err != nil {
-		return nil, err
-	}
-	if code != http.StatusOK {
-		return nil, fmt.Errorf("Failed to get default server monitor config (%s, status %d)", url, code)
-	}
-	config := &mmServer.Config{}
-	if err := json.Unmarshal(data, config); err != nil {
-		return nil, err
-	}
-	config.Service = "server"
-	config.InstanceId = si.Id
-
-	bytes, err := json.Marshal(config)
-	if err != nil {
-		return nil, err
-	}
-	agentConfig := &proto.AgentConfig{
-		InternalService: "mm",
-		ExternalService: proto.ServiceInstance{
-			Service:    "server",
-			InstanceId: si.Id,
-		},
-		Config:  string(bytes),
-		Running: true,
-	}
-	return agentConfig, nil
-}
-
-func (i *Installer) getMmMySQLConfig(mi *proto.MySQLInstance) (*proto.AgentConfig, error) {
-	url := pct.URL(i.agentConfig.ApiHostname, "/configs/mm/default-mysql")
-	code, data, err := i.api.Get(i.agentConfig.ApiKey, url)
-	if i.flags.Bool["debug"] {
-		log.Printf("code=%d\n", code)
-		log.Printf("err=%s\n", err)
-	}
-	if err != nil {
-		return nil, err
-	}
-	if code != http.StatusOK {
-		return nil, fmt.Errorf("Failed to get default MySQL monitor config (%s, status %d)", url, code)
-	}
-	config := &mmMySQL.Config{}
-	if err := json.Unmarshal(data, config); err != nil {
-		return nil, err
-	}
-	config.Service = "mysql"
-	config.InstanceId = mi.Id
-
-	bytes, err := json.Marshal(config)
-	if err != nil {
-		return nil, err
-	}
-	agentConfig := &proto.AgentConfig{
-		InternalService: "mm",
-		ExternalService: proto.ServiceInstance{
-			Service:    "mysql",
-			InstanceId: mi.Id,
-		},
-		Config:  string(bytes),
-		Running: true,
-	}
-	return agentConfig, nil
-}
-
-func (i *Installer) getSysconfigMySQLConfig(mi *proto.MySQLInstance) (*proto.AgentConfig, error) {
-	url := pct.URL(i.agentConfig.ApiHostname, "/configs/sysconfig/default-mysql")
-	code, data, err := i.api.Get(i.agentConfig.ApiKey, url)
-	if i.flags.Bool["debug"] {
-		log.Printf("code=%d\n", code)
-		log.Printf("err=%s\n", err)
-	}
-	if err != nil {
-		return nil, err
-	}
-	if code != http.StatusOK {
-		return nil, fmt.Errorf("Failed to get default MySQL sysconfig config (%s, status %d)", url, code)
-	}
-	config := &sysconfigMySQL.Config{}
-	if err := json.Unmarshal(data, config); err != nil {
-		return nil, err
-	}
-	config.Service = "mysql"
-	config.InstanceId = mi.Id
-
-	bytes, err := json.Marshal(config)
-	if err != nil {
-		return nil, err
-	}
-	agentConfig := &proto.AgentConfig{
-		InternalService: "sysconfig",
-		ExternalService: proto.ServiceInstance{
-			Service:    "mysql",
-			InstanceId: mi.Id,
-		},
-		Config:  string(bytes),
-		Running: true,
-	}
-	return agentConfig, nil
-}
-
-func (i *Installer) getQanConfig(mi *proto.MySQLInstance) (*proto.AgentConfig, error) {
-	url := pct.URL(i.agentConfig.ApiHostname, "/configs/qan/default")
-	code, data, err := i.api.Get(i.agentConfig.ApiKey, url)
-	if i.flags.Bool["debug"] {
-		log.Printf("code=%d\n", code)
-		log.Printf("err=%s\n", err)
-	}
-	if err != nil {
-		return nil, err
-	}
-	if code != http.StatusOK {
-		return nil, fmt.Errorf("Failed to get default Query Analytics config (%s, status %d)", url, code)
-	}
-	config := &qan.Config{}
-	if err := json.Unmarshal(data, config); err != nil {
-		return nil, err
-	}
-	config.Service = "mysql"
-	config.InstanceId = mi.Id
-
-	bytes, err := json.Marshal(config)
-	if err != nil {
-		return nil, err
-	}
-	agentConfig := &proto.AgentConfig{
-		InternalService: "qan",
-		ExternalService: proto.ServiceInstance{
-			Service:    "mysql",
-			InstanceId: mi.Id,
-		},
-		Config:  string(bytes),
-		Running: true,
-	}
-	return agentConfig, nil
-}
 
 func (i *Installer) writeInstances(si *proto.ServerInstance, mi *proto.MySQLInstance) error {
 	// We could write the instance structs directly, but this is the job of an
 	// instance repo and it's easy enough to create one, so do the right thing.
 	logChan := make(chan *proto.LogEntry, 100)
 	logger := pct.NewLogger(logChan, "instance-repo")
-	repo := instance.NewRepo(logger, pct.Basedir.Dir("config"), i.api)
+	repo := instance.NewRepo(logger, pct.Basedir.Dir("config"), i.apiConnector)
 	if si != nil {
 		bytes, err := json.Marshal(si)
 		if err != nil {
